+++ Online-Shop/Collections/Node.cs
@@ -4,14 +4,14 @@ using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
 
-namespace Online_Shop
+namespace Online_Shop.Collections
 {
     public class Node<T>
     {
         public T Data { get; set; }
-        public  Node<T> Next { get; set; }
+        public Node<T> Next { get; set; }
 
-        public Node(T data, Node<T> next) 
+        public Node(T data, Node<T> next)
         {
             Data = data;
             Next = next;
