+++ Online-Shop/Core/DiscountOrder.cs
@@ -1,23 +1,17 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-
-namespace Online_Shop
+﻿namespace Online_Shop.Core
 {
-    internal class DiscountOrder: Order
+    internal class DiscountOrder : Order
     {
-        public float Discount {  get; set; }
+        public float Discount { get; set; }
 
         private float _price;
         public new float Price
         {
-            get {  return _price; }
+            get { return _price; }
             set { _price = value * Discount / 100; }
         }
 
-        public DiscountOrder(string name, long phone, float price, string address, float discount): base(name, phone, price, address)
+        public DiscountOrder(string name, long phone, float price, string address, float discount) : base(name, phone, price, address)
         {
             Discount = discount;
         }
