+++ Online-Shop/Core/VIPOrder.cs
@@ -1,17 +1,10 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Security.Cryptography.X509Certificates;
-using System.Text;
-using System.Threading.Tasks;
-
-namespace Online_Shop
+﻿namespace Online_Shop.Core
 {
-    internal class VIPOrder: Order
+    internal class VIPOrder : Order
     {
         public string Present { get; set; }
 
-        public VIPOrder(string name, long phone, float price, string address, string present): base(name, phone, price, address) 
+        public VIPOrder(string name, long phone, float price, string address, string present) : base(name, phone, price, address)
         {
             Present = present;
         }
