+++ Online-Shop/Program.cs
@@ -8,9 +8,9 @@ Task 5
 6*. Вывести на экран название товара, наиболее часто встречающееся в списке заказов
 */
 
-using Online_Shop;
-using System;
-using System.Collections.Generic;
+using Online_Shop.Collections;
+using Online_Shop.Core;
+
 
 Order[] orders = { new("soundbar LG SJ3", 80175550011, 770, "Minsk, Kirova str. 14, apt. 25"),
                    new("refrigerator ATLANT 4626", 375173335544, 1629.55f, "Fanipol, Komsomolskaya str. 7, apt. 43"),
